Allow specifying ip to connect to
A server in the config is specified by a hostname. A hostname can resolve to
multiple ip addresses. Enable specifying an ip address in the config. When set
the ip address is used when creating the tcp connection, while the server name
continues to be used for TLS etc. Adding the `ip` option enables workflows such
as pre-selecting the most optimal ip to connect to (e.g. based on latency),
while still validating certs etc.